refactor: split REST sync anchors
