+++ .sampo/changesets/1014-split-rest-sync-anchors.md
@@ -1,0 +1,6 @@
+---
+npm/@wp-typia/project-tools: patch
+npm/wp-typia: patch
+---
+
+Split REST sync-rest anchor patching into contract, resource, and shared helper modules while preserving compatibility exports.
